refactor(desktop): 确认卡提示生产接线改顶层静态 import
codex review P1:main 进程禁止运行时动态 import()(architecture-invariants
§2),desktopConfirmNotice 的两处 import() 违规。

拆法:纯逻辑留在 desktopConfirmNotice.ts(零 electron/db 依赖,单测直接引);
createFeishuDesktopConfirmNotifier 挪进新的 desktopConfirmNoticeWiring.ts,
bindingStore / feishuIm / localDb 全部顶层静态 import。host 与 binding 本来
就在 main 的静态依赖图里(bootstrap-electron → im/index → host),原先
「避免模块加载副作用」的动态加载理由不成立。
